Changed ETAX PV to be a double

+++ BLDApp/src/BLDMCast.h
@@ -189,7 +189,7 @@ extern IOSCANPVT    bldIoscan;         /* Trigger full-rate EPICS records */
 #define BLDMCAST_DST_PORT	10148
 
 // For use with the ETAX PV. See devBLDScalars.c
-long etaxPV;
+double etaxPV;
 
 #ifdef __cplusplus
 }
